requirements.txt
anthropic
chromadb
sentence-transformers
requests
httpx
python-dotenv
apscheduler
python-telegram-bot
beautifulsoup4